refactored folders and user schema code
